requirements.txt
flask==1.1.2
Flask-Admin==1.5.7
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
GeoAlchemy2==0.8.4
geos==0.2.2
Shapely==1.7.1